refactor:reorganize components and functions in utils
