+++ src/components/Main/HotelCard/HotelCard.jsx
@@ -1,8 +1,10 @@
+//Dependencies
 import React from 'react';
-import {getPriceSigns, getBackgroundImg} from '../../../assets/utils.js'
+//Utils
+import {getPriceSigns, preventScroll, restartScroll} from '../../../assets/utils.js'
 
 const HotelCard = ({name, country, rooms, price, description, photo}) => {
-    
+    //Handler
     const handleTap = (e) => {
         const hotelCard = e.target.parentElement;
         hotelCard.classList.toggle('Expanded');
@@ -35,16 +37,3 @@ const HotelCard = ({name, country, rooms, price, description, photo}) => {
 }
 
 export default HotelCard;
-
-function preventScroll(){
-    window.scrollTo(0, 0);
-    window.addEventListener('scroll', scrollToTop);
-}
-
-function scrollToTop(){
-    window.scrollTo(0, 0);
-}
-
-function restartScroll(){
-    window.removeEventListener('scroll', scrollToTop);
-}
